Feat/345657

+++ LudenWebAPI/Application/Abstractions/Interfaces/Repository/IBillRepository.cs
@@ -4,6 +4,6 @@ namespace Application.Abstractions.Interfaces.Repository
 {
     public interface IBillRepository : IGenericRepository<Bill>
     {
-        Task<IEnumerable<Bill>> GetBillsByUserIdAsync(int userId);
+        Task<IEnumerable<Bill>> GetBillsByUserIdAsync(ulong userId);
     }
 }
